Update sample Ford Escape config
Updated Ford Escape config to use new docker compose structure as well as new
folder structure for CARMA configuration. This was in response to issue #662
and should resolve the underyling issue, though it's still not expected that
the Ford Escape config is fully functional (as it is not a fully supported
vehicle for CARMA3) or that it would be feasible to run the Ford Escape config
on a development PC without the Ford Escape hardware or IP address config.