Improve through scenario (#331)
<!-- Thanks for the contribution, this is awesome. -->

# PR Details
## Description

<!--- Describe your changes in detail -->
Added a new scenario-runner scenario option to use for the "through"
testing.
This is because we need longer time for the VRU to be blocked, so I
moved the truck to right more. And also moved the pedestrian to the edge
of the crosswalk because the current options for VRU speed (0.9-1.2 m/s)
is not enough to produce collision.

Added carma-platform's osm map where the map's speed is all changed to
35mph max. The osm map in MAP folder has many different speed variations
including inf and has no elevations so visualization had little offset.
I tried using this new map for the intersection_model/tsc_service but it
is not working (tsc_service never activates in time).
Currently I don't have time to investigate what is the difference and
why. So I am just including both maps.

It also has the configuration for collision radius increase that has
been used for the first left turn scenario. It was really meant for
through scenario, but in efforts to keep the parameters same the data
was collected with that number

## Related Issue
https://usdot-carma.atlassian.net/browse/CDAR-662
<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an
issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps
to reproduce -->
<!--- Please link to the issue here: -->

## Motivation and Context
vru verifiacation test prep
<!--- Why is this change required? What problem does it solve? -->

## How Has This Been Tested?
sim pc 2
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [ ] Defect fix (non-breaking change that fixes an issue)
- [X] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that cause existing functionality
to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [X] I have added any new packages to the sonar-scanner.properties file
- [ ] My change requires a change to the documentation.
- [X] I have updated the documentation accordingly.
- [X] I have read the **CONTRIBUTING** document.
[CARMA Contributing
Guide](https://github.com/usdot-fhwa-stol/carma-platform/blob/develop/Contributing.md)
- [X] I have added tests to cover my changes.
- [X] All new and existing tests passed.